requirements.txt
discord.py
huggingface_hub
python-dotenv